Move to Hash.Swift.
